Move the X module bits into xf86Wacom.c
Split preinit and uninit into two parts each, one for the
driver-specific stuff, one for what is purely X specific stuff.
